refactored genomics.py
