+++ src/nplinker/genomics/MiBIGBGC.py
@@ -1,0 +1,14 @@
+from .BGC import BGC
+
+
+class MiBIGBGC(BGC):
+
+    def __init__(self, id, strain, name, product_prediction):
+        super().__init__(id, strain, name, product_prediction)
+
+    def __eq__(self, other):
+        return self.id == other.id
+
+    def __hash__(self):
+        return self.id
+
